fix(entity-generator): allow arbitrary class and prop names as identifiers
AbstractNamingStrategy is updated to handle such conflicts, and inherently,
any naming strategy that inherits from it.

Any table that starts with a character that is invalid as a JS start character (not just digits)
gets prefixed with "E".
Characters that are not valid in any position as a class identifier (or the character "$")
get turned into code points, prefixed with $, making them valid (though inconvenient to type)
identifiers and file names.

Prop names that conflict with the prototype of Reference (which all loaded entities inherit from)
get prefixed with "$", thus allowing the safe generation from
a column named "get", "load", "$", etc. (declared as props "$get", "$load", "$$", etc.).

Other prop names that are otherwise not valid identifiers get string quoted.
This is a SourceFile/EntitySchemaSourceFile feature, not a naming strategy feature,
so it will get applied regardless of naming strategy.

Also made the imports of types tied to the "entity" prop option, rather than the prop declaration.
This ensures proper imports when using mapToPk in extension hooks.

Also moved the schema de-duplication code to happen before bi-directional and many-to-many
detection, to ensure less work and error prone-ness for the other stages in updating references.

Also fixed quoting of strings in generated columns.